requirements.txt
SQLAlchemy~=2.0.29
uvicorn~=0.29.0
fastapi~=0.110.1
pydantic~=2.6.4
starlette~=0.37.2
pytest~=8.1.1
httpx~=0.27.0
pytest-asyncio~=0.23.6
aiosqlite~=0.20.0
PyYAML~=6.0.2